chore: extract getTemplateVersionNumber into shared utils to reduce duplication (PIN-9268)
